requirements.txt
# Core ML
<<<<<<< HEAD
# GPU (NVIDIA): install CUDA PyTorch first — plain `pip install torch` is CPU-only on Windows.
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
# Then: pip install -r requirements.txt
torch>=2.6.0
=======
torch>=2.0.0
>>>>>>> f28ad62 (My local changes before pulling)
transformers==4.44.2
datasets==2.19.1
accelerate==0.33.0
peft==0.11.1
bitsandbytes==0.42.0


# RAG & Embeddings
sentence-transformers==2.3.1
faiss-cpu

# Data Processing
pandas==2.1.4
# accelerate 0.33.x requires numpy<2; keep 1.x for the full ML stack on Windows.
numpy>=1.26.4,<2.0
pdfplumber==0.10.3
beautifulsoup4==4.12.2

# API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
SQLAlchemy>=2.0.29
passlib>=1.7.4
pymongo>=4.8.0
dnspython>=2.6.1
python-multipart

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
PyYAML==6.0.1

# Document Processing
python-docx==1.1.0
docxtpl==0.20.2
PyMuPDF>=1.24.0
docx2pdf>=0.1.8
scikit-learn==1.3.2